google-trends-analysis/requirements.txt
pytrends>=4.9.2
pandas
matplotlib
python-bidi
jsonschema
